Refactor into packages for core, extensions and maestro
