+++ benchmarks/test_benchmark_maze_loopfree.py
@@ -3,10 +3,10 @@ from itertools import product
 
 import pytest
 
-from cosy.specification_builder import SpecificationBuilder
-from cosy.synthesizer import Specification, Synthesizer
-from cosy.tree import Tree
-from cosy.types import Constructor, DataGroup, Literal, Type, Var
+from cosy.core.specification_builder import SpecificationBuilder
+from cosy.core.synthesizer import Specification, Synthesizer
+from cosy.core.tree import Tree
+from cosy.core.types import Constructor, DataGroup, Literal, Type, Var
 
 
 def is_free(pos: tuple[int, int]) -> bool:
