+++ examples/example_constraints_inline.py
@@ -6,8 +6,8 @@ Demonstrates constraints in CoSy, highlighting how the Component class can be us
 import re
 
 from cosy import Maestro
-from cosy.specification_builder import SpecificationBuilder
-from cosy.types import Constructor, Group, Literal, Type, Var
+from cosy.core.specification_builder import SpecificationBuilder
+from cosy.core.types import Constructor, Group, Literal, Type, Var
 
 
 def main():
