--- src/cosy/__init__.py
@@ -1,19 +1,0 @@
-__all__ = [
-    "Arrow",
-    "Constructor",
-    "Intersection",
-    "Literal",
-    "Maestro",
-    "Omega",
-    "SpecificationBuilder",
-    "Subtypes",
-    "Synthesizer",
-    "Type",
-    "Var",
-]
-
-from cosy.maestro import Maestro
-from cosy.specification_builder import SpecificationBuilder
-from cosy.subtypes import Subtypes
-from cosy.synthesizer import Synthesizer
-from cosy.types import Arrow, Constructor, Intersection, Literal, Omega, Type, Var
