+++ src/cosy/core/__init__.py
@@ -1,0 +1,17 @@
+__all__ = [
+    "Arrow",
+    "Constructor",
+    "Intersection",
+    "Literal",
+    "Omega",
+    "SpecificationBuilder",
+    "Subtypes",
+    "Synthesizer",
+    "Type",
+    "Var",
+]
+
+from cosy.core.specification_builder import SpecificationBuilder
+from cosy.core.subtypes import Subtypes
+from cosy.core.synthesizer import Synthesizer
+from cosy.core.types import Arrow, Constructor, Intersection, Literal, Omega, Type, Var
