+++ src/cosy/core/combinatorics.py
