+++ src/cosy/core/tree.py
