+++ src/cosy/core/types.py
