+++ src/cosy/extensions/__init__.py
@@ -1,0 +1,6 @@
+__all__ = [
+    "MaestroSolutions",
+    "Solutions",
+]
+
+from cosy.extensions.solutions import MaestroSolutions, Solutions
