+++ src/cosy/maestro/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "Maestro",
+]
+
+from cosy.maestro.maestro import Maestro
